# Documentation/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BetMachine",
    "version": "1.0"
  },
  "paths": {
    "/api/Match": {
      "get": {
        "tags": [
          "Match"
        ],
        "operationId": "GetMatch",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDtoServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDtoServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDtoServiceResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Match"
        ],
        "operationId": "AddMatch",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Match"
        ],
        "operationId": "UpdateMatch",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MatchDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Match"
        ],
        "operationId": "DeleteMatch",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/MatchOdd": {
      "get": {
        "tags": [
          "MatchOdd"
        ],
        "operationId": "GetMatchOdd",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOddServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOddServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOddServiceResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MatchOdd"
        ],
        "operationId": "AddMatchOdd",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchOddDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchOddDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchOddDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddMatchOddDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MatchOdd"
        ],
        "operationId": "UpdateMatchOdd",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MatchOddDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchOddDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchOddDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MatchOddDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MatchOdd"
        ],
        "operationId": "DeleteMatchOdd",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanServiceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddMatchDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "string",
            "format": "date-time"
          },
          "matchTime": {
            "type": "string",
            "example": "00:00:00"
          },
          "teamA": {
            "type": "string",
            "nullable": true
          },
          "teamB": {
            "type": "string",
            "nullable": true
          },
          "sport": {
            "$ref": "#/components/schemas/Sports"
          }
        },
        "additionalProperties": false
      },
      "AddMatchOddDto": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "integer",
            "format": "int32"
          },
          "specifier": {
            "type": "string",
            "nullable": true
          },
          "odd": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "BooleanServiceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Int32ServiceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "integer",
            "format": "int32"
          },
          "success": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Match": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "string",
            "format": "date-time"
          },
          "matchTime": {
            "type": "string",
            "example": "00:00:00"
          },
          "teamA": {
            "type": "string",
            "nullable": true
          },
          "teamB": {
            "type": "string",
            "nullable": true
          },
          "sport": {
            "$ref": "#/components/schemas/Sports"
          },
          "matchOdds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchOdd"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MatchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "matchDate": {
            "type": "string",
            "format": "date-time"
          },
          "matchTime": {
            "type": "string",
            "example": "00:00:00"
          },
          "teamA": {
            "type": "string",
            "nullable": true
          },
          "teamB": {
            "type": "string",
            "nullable": true
          },
          "sport": {
            "$ref": "#/components/schemas/Sports"
          }
        },
        "additionalProperties": false
      },
      "MatchDtoServiceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MatchDto"
          },
          "success": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MatchOdd": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "matchId": {
            "type": "integer",
            "format": "int32"
          },
          "specifier": {
            "type": "string",
            "nullable": true
          },
          "odd": {
            "type": "number",
            "format": "double"
          },
          "match": {
            "$ref": "#/components/schemas/Match"
          }
        },
        "additionalProperties": false
      },
      "MatchOddDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "matchId": {
            "type": "integer",
            "format": "int32"
          },
          "specifier": {
            "type": "string",
            "nullable": true
          },
          "odd": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "MatchOddServiceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MatchOdd"
          },
          "success": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "extraDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Sports": {
        "enum": [
          "None",
          "Football",
          "Basketball"
        ],
        "type": "string"
      }
    }
  }
}